search: Add #searchbox_legacy selector to #searchbox rules in media.scss.
This is a fixup for 8c5a77d17ae17bc6221afd214200c5c8f4f6eaeb.
While copying the the rules of `#searchbox` to `#searchbox_legacy`
in the search pills feature, the existing `#searchbox` rules were
missed in the conversion.
`#searchbox_legacy` has been added beside `#searchbox` in `media.scss`
instead of replacing that as both of them need those rules for the
mobile view.