pills: Add `flex-wrap: wrap` to pill container for all pills.
Fixes #10059.
In e76d046adc8c030f0a7ae3ff12b91960b76e6b5a,
`display: inline-flex` was added to `.pill-container` but
`flex-wrap: wrap` was missing which forced overflow pills to be on
one line and made the pill text overflow vertically. This was not
observed in composebox pills as `.pm_recipient .pill-container`
already had a `flex-wrap: wrap` rule which has been removed in this
commit to avoid duplication.