hotkeys: Use canonical name for reacting with `+` hotkey.
This is sort of a temporary fix to bring the state back to how it
was in commit: a907fe3a5ee6213340f47eca1c3b170488ef3f18. However,
long-term we will need to fix our local echo feature to do merging
of names just like we do on backend.