license: Move license application notice from LICENSE to NOTICE.
The boilerplate here is supposed to be example of how to apply the Apache
license to files in the project, and is not intended to be applied to the
LICENSE file itself. See https://github.com/benbalter/licensee/issues/103
for some discussion of this.

Moved the text from the LICENSE file to a NOTICE file, with one wording
change: "this file" -> "this project".