emails: Fix wrongly converted hex colors.
It looked like I got some hex values wrong during the
hsl to hex conversions. I used the built in vscode color
converter for this so not really sure how this happened.
The colors are now equalent to the original hex colors
before the conversion to hsl was made in
3de2953b0318f60e4cd26c907602aea8b9584128