docs: Update dependencies for docs.
This updates recommonmark so that relative links to .md files will
correctly be converted to links to html pages by Sphinx/ReadTheDocs.
This was repoprted in https://github.com/rtfd/recommonmark/issues/89

This will allow us to in a future commit use relative links to .md
files so all of the hyperlinks work both on ReadTheDocs and on GitHub.